Make it possible to specify the path of a pkg in `include_xxx` config
Add a .path attribute to Autoimport and import the pkg from there if
given. This allows to keep the pkg defining the include/libraries
finding functions for a required library with the project and have it
work even when installing it in editable mode.

Note that the build process still needs to have write access to the
source tree of the project for this to work.